Fix seeding and add seeding with strings
Before, the seeding did not work when the user specified a seed
via parameter. This is because the seed was never an int but a
string when entered as ID2T parameter.

Now: The original functionality persists, i.e. if the seed comes
as int, it is processed as int. If the seed is a string, but
contains only digits, it is converted to an int. Finally, if the
seed is a string that contains non-digit characters, the string
is hashed to an int, which is then used as seed.